requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Authlib==1.3.2
PyJWT==2.9.0
Werkzeug==3.0.4
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
python-dotenv==1.0.1
requests==2.32.3
celery[redis]==5.4.0
Flask-Mail==0.10.0
Flask-Bcrypt==1.0.1
redis==5.1.1
pillow==11.0.0
apscheduler==3.10.4
flask-sse==1.0.0